Update main door Jonathan

+++ src/main/java/be/intec/themarujohyperblog/service/PostServiceImpl.java
@@ -151,7 +151,7 @@ public class PostServiceImpl implements PostService {
     }
 
 
-    public int countPosts() {
+    public long countPosts() {
         return (int) postRepository.count();
     }
